Add empty Page
When creating a new Pdf document there are no pages.
Trying to sign such document ends with an error,
that the document needs at least one page.

It does not really make sense to have the document without pages,
and if someone desires such document,
he still has the option to delete the pages.