Remove content instruction observers
This functionality was used before, when every change to the content
instruction should have triggered content stream data recalculation.

Today this requires an explicit call,
so this subscriptions and monitoring is not needed.

Not only it requires more memory,
but it also slows the application down.

Even though I do not have the benchmarks yet,
I feel confident this is not needed.