FIX: XRefStm dual entry
Files with XRefStm do have two different entries for some objects.
The implementation did not handle this, meaning this would read the same
object multiple times from disk and in case there are changes,
the behavior was unspecified.

With this change all of such entries would the the same object in memory