NEW TEST DOCUMENT EDIT
Document edit modifies the first page contents to verify,
that the changes are being persisted.

There was recently an issue, where for file with XRefStm
such changes were not persisted.
Even though this does does not fully cover this specific scenario
it is a nice extension to the test suite.