FIX: Move xref entry initialization
Until now, all of the used xref entries were
initialized during the file initialization.
This slows down the initialization, but moreover it has issues with
the files, where the stream size is in form of indirect reference.

The logic was moved to the file GetIndirectObjectInternal,
which maybe called also during file initialization,
however not for all objects by default.

Under some circumstances, when the external application
asks for all xref entries and tries to get the object reference,
it could point out to a different object,
however in such case the source file is heavily corrupted
even though it could be loaded.

Such scenarios should be handled within the external application.