FIX: Xref unordered_set
When investigating the differences using set vs unoredered set on
linux GCC, I went deeper in this and even though I have not found
the root cause for this, under some circumstances
we are changing the xref entry object number.
This should not happen and for the time being I have not changed the key
in the xref table map, however we are reinserting
all the elements in such cases.
It is currently only present in one single document,
so slowing down initialization for all document seems to be overkill.