FIX: RemoveEncryption issues
Removing encryption from file did not mark the streams as dirty
and the entire content was copied from the original file.
For this reason a new flag IsEncryptionDirty was added to file interface.
This will be covered in the tests later.

There was one private file that contained the permissions 4294967232,
which is outside of int32 and caused the application crash.

The decryption tool is now able to save the file to a disk.