TWEAK: Change linux-x64 and arm64 to rockylinux.9
For some reason the rockylinux.8 embeds a dependency on pthread and libdl.
I was not able to load the binary on ubuntu 24.04 for this reason,
where compiling with rockylinux.9 resolves this issue.

This probably means, that the deployment on
rockylinux.8 and similar will not work.

For the time being I do accept this impact as we need to move on.