NEW: sanity-check.yml (#12)
* NEW: sanity-check.yml

Add basic build and test for all 3 platforms.
arm platforms are missing for the time being as we need to decide
if we go with the docker image approach or we use actual arm VMs.

* FIX: Sanity check

Windows image cannot call make as this is UNIX utility.
Let's use cmake and ctest, which should be cross-platform.

* FIX: Sanity check

Syntax issue for build type

* FIX: Sanity check

Generate project files prior to running build.
Add macos-latest to see where we are with arm mac.

* TWEAK: vcpkg metrics

Disable vcpkg metrics when building in pipelines