Vcpkg portfile (#31)
* NEW: vcpkg portfile

Preparation to build vcpkg packages.
vcpkg bootstrapping is causing issues,
when being used as a submodule in other projects.
We need to move the responsibility to the top project.

* FIX: cmake bool

* FIX: VANILLAPDF_STANDALONE

It should be option and not a cache flag.

* TWEAK: vcpkg portfile

Add options to disable tests and benchmark due to dependencies.

* FIX: natvis linking

Path to natvis is dependent on the working folder.
fatal error LNK1104: cannot open file 'public.natvis'

* FIX: natvis linking

Try to use VANILLAPDF_SOLUTION_SOURCE_DIR as there is still linking problem.

* TWEAK: Change doc url

* FIX: natvis linking

* FIX: cmake exports

Missing export and config files to be consumed by the parent project.

* FIX: metadata folder

License and readme should be in the share folder and not root.
vcpkg issues a warning, so let's be compliant.

* NEW: vcpkg usage

* FIX: portfile

Additional cleanup for debug share.
Include NOTICE.md in the root as well.

* Cleanup

* FIX: Missing imports

When consuming the package public headers are not available.
We need to publish this as target property.

* NEW: BUILD_SHARED_LIBS flag

Add support for universal BUILD_SHARED_LIBS flag.

* TWEAK: Enable PDB

Generate PDB files by default as this is standard behavior.
All other libs - zlib, openssl, libjpeg - produce PDB
both in debug and release.

* TWEAK: Install PDB

Include PDB files in the install package.

* FIX: PDB only for MSVC

Installing PDBs only works for MSVC, remove for other platforms.

* FIX: CodeQL failure

We have removed fallback for CMAKE_TOOLCHAIN_FILE,
which can cause issues.
The fallback is actually working
since we are setting this before the first call to project().