Improve logging (#35)
* TWEAK: Log and throw

This is the most often happening error, that would be nice to see
immediately in the console log to see, what is going on.
If we do not log this, we need to parse the error codes and extract text.

* TWEAK: More LOG_ERROR_AND_THROW_GENERAL

Cleanup of parameters, reduce verboseness of stringstream.

* TWEAK: Disable MSVC warning

MSVC started warning recently about unused parameters.
This is not a real issue, so let's suppress this warning.