FIX: Missing dependency linking
When building static library the parent project is not able to
resolve some internal targets, such as zlib or openssl.
It works when the parent project calls find_package,
however this add burden to their maintenance, which is unprofessional.
We should be able to fix this using find_dependency inside vcpkg install scripts.