Static library (#36)
* NEW: Support for static lib

Static or shared lib is being configured using BUILD_SHARED_LIBS.
https://cmake.org/cmake/help/latest/variable/BUILD_SHARED_LIBS.html

* FIX: portfile static lib support

Removed vcpkg_check_linkage(ONLY_DYNAMIC_LIBRARY).

* FIX: Dependency leaking

Rework target_link_libraries, target_include_directories and target_compile_definitions.

* FIX: Missing dependency linking

When building static library the parent project is not able to
resolve some internal targets, such as zlib or openssl.
It works when the parent project calls find_package,
however this add burden to their maintenance, which is unprofessional.
We should be able to fix this using find_dependency inside vcpkg install scripts.

* FIX: Install targets

@PACKAGE_INIT@ was not properly expanded,
due to error, where some of the paths were incorrectly set.
Let's try to fix vcpkg install script relocation addresses.

* FIX: vcpkg dependencies

Resolution of the dependencies is done when
processing vanillapdfTargets.cmake and not it's execution.
We need to move dependency loading and resolution to another file,
that will be loaded before processing vanillapdfTargets.cmake.

* FIX: vcpkg install cleanup

Some duplicit targets were added in the last commit,
that broke the build script.

* TWEAK: Add import guard

In case of duplicit import of the project,
the dependencies would be added multiple times.
We have added a guard,
that only allows the dependencies to be added once.

* FIX: pthread on MSVC

spdlog requires thread library for multi-thread support.
MSVC compiler is having some issues linking this on windows
and has to explicitly told not to prefer pthread.

* FIX: Missing openjp

* FIX: pthread issue

When compiling x64-windows-static-md the issue
with linking pthread.lib happens again.
It seems that there is another flag controlling this behavior,
that needs to be modified in this build configuration.

* FIX: pthread link

The resolution of pthread on windows is very complex and having
these flags in the vanillapdf subproject might be too late.
Let's try to put it in the root project.

* FIX: pthread link

Still not working, trying to move it before vcpkg initialization.

* FIX: pthread link

I think the issue is connected in the process of
how vcpkg resolves the dependencies.
Let's include it in the installation process,
so it is defined as early as possible.

* TWEAK: Revert pthread changes

Let's go back to basics, the x64-windows-static works.
The x64-windows-static-md is having issues,
however let's have a separate PR for this.

---------

Signed-off-by: Juraj Zikmund <jur.zikmund@gmail.com>