FIX: Broken debug
Assertions triggered in debug builds, however the assertion in that place is not correct.
Assertion points to developer bug, where in this case the issue was with the input file.
We are logging warning instead and there is GitHub issue addressing the issue.

The decision to move on has to be made, since the current setup in release mode does not do anything.
We are improving the behavior even though not resolving the issue entirely.

https://github.com/vanillapdf/vanillapdf/issues/37