TWEAK: yml updates
Disable ubuntu 22.04-arm as we are not able to build docker image.

Disable cache for nightly and nuget - we should not share the cache between workflows.
Since this is running once a day/week we don't necessarily need cache.
Let's keep it only for sanity, which is being run all the time.

Disable vcpkg manifest for vanillapdf core - we don't need any dependencies