Code coverage (#38)
* FIX: Broken debug

Assertions triggered in debug builds, however the assertion in that place is not correct.
Assertion points to developer bug, where in this case the issue was with the input file.
We are logging warning instead and there is GitHub issue addressing the issue.

The decision to move on has to be made, since the current setup in release mode does not do anything.
We are improving the behavior even though not resolving the issue entirely.

https://github.com/vanillapdf/vanillapdf/issues/37

* TWEAK: Log and throw

Slight readability and performance improvement.
We also log error and not only throw the exception.

* NEW: Code coverage

yml, compile and link options

* FIX: Permissions

Codecov wanted to update statuses and comment on PR.
Let's add explicit permissions allowing it to do so.

* FIX: Missing token

Codecov profile was activated, secret added.

* FIX: Coverage report

Current coverage report actually shows the lines in the test projects.
What we want to have visibility on the library about the coverage.
Let's enable coverage report for vanillapdf only.

* TWEAK: Coverage badge

Improve readme, add new badge for code coverage

* NEW: Add codecov.yml config