Linux arm64 (#40)
* TWEAK: Enable linux arm64

It should be available in public beta:
https://github.com/actions/partner-runner-images

* FIX: Cannot find GCC

aarch64-linux-gnu-gcc was used for cross-compilation.
Since we are running natively on arm, we should be good with gcc.

* FIX: arm64 warning

On Linux arm64 the Buffer::value_type seems to be unsigned char,
which is causing warning, that the condition is always true.
We should not rely on Buffer::value_type is unsigned, so let's fix it.

* TWEAK: Enable arm64 in CI

Enable linux-arm64 also for nightly check and sanity.
Increase nightly check to be done only once a week - sanity should be good enough for most cases.

* FIX: arm64 os

Select specific runner os to be used for each CI configuration.

* TWEAK: yml updates

Disable ubuntu 22.04-arm as we are not able to build docker image.

Disable cache for nightly and nuget - we should not share the cache between workflows.
Since this is running once a day/week we don't necessarily need cache.
Let's keep it only for sanity, which is being run all the time.

Disable vcpkg manifest for vanillapdf core - we don't need any dependencies

* Revert "TWEAK: yml updates"

This reverts commit d351535c2de2291ea5051203e47b674f366da240.

* TWEAK: Disable ubuntu 22.04 arm

There is a QEMU bug in WSL, that cannot build ubuntu 22.04 with arm emulation.
We do have ubuntu 24.04-arm, so we should be good for now.