TWEAK: Workflow trigger on change
When there is a change in the workflow file we should ensure,
that it is safe to be executed.

For sanity this happens in all cases, however the nuget and nightly check are only manual or scheduled.
Let's ensure, that the safe branch can execute them in case they were changed in the PR.