Workflow fixes (#41)
* TWEAK: Add coverage concurrency

When there are multiple pushes to main branch I don't see the reason to let coverage report to finish.
It slows down the entire pipeline process and only the last commit is actually relevant.
All of the other check should be done on the branch.

* FIX: Compiler flags

Debug configuration for MSVC was not properly set, let's fix it.

* NEW: Debug/Release check

Add both debug and release check into sanity,
to increase the overall stability of the main branch.

* FIX: macOS quiet

When installing the already installed package using brew,
there is a warning that the package is already installed.
This is not a bug, so let's suppress this annotation.

* TWEAK: Workflow trigger on change

When there is a change in the workflow file we should ensure,
that it is safe to be executed.

For sanity this happens in all cases, however the nuget and nightly check are only manual or scheduled.
Let's ensure, that the safe branch can execute them in case they were changed in the PR.

* TWEAK: CI workflows

Add debug/release to nightly check workflow.
Update nuget and sanity job names.
Add max parallel for sanity check.

* FIX: Nuget build

Missing --config for windows MSVC package, that triggers both debug and release.
MSVC is multi-config generator, where ninja is not.

* FIX: Disable windows debug

win-x86 is having some serious issues with one document in debug mode,
taking more than 1500 seconds to finish, which times out the test.
I have tried to do some optimizations in the vanillapdf to reduce this,
however I wasn't able to reach desired performance yet.

Since debug mode is not necessary critical component and
we have debug enabled on other platforms,
let's remove this from the windows configuration.