FIX: NuGet package location
There is a complex issue with picking up correct packages after build.
MSVC has a different structure than ninja and we need to merge them.
For the time being packaging is done through CI, so we can rely on this.

Additionally I have removed vanillapdf.lib, which was used to aid
semi-dynamic linking, when the package was consumed by native project.
The issues keeping up the native targeting are high and for this reason
we now rely mainly on the github clone and vcpkg.