Switch NuGet packaging to dotnet pack (#50)
* Remove Nuspec packaging

* restore before dotnet pack

* Fix newline in NuGet targets template

* Document NuGet restore step

* Install dotnet in build workflow

* Improve NuGet packaging workflow

* Install ICU library in build

* Fix core packaging job dependencies

* Fix NuGet artifact upload and silence build warning

* Use netstandard TFM for NuGet projects

* Remove native placeholder

* Generate dependency groups for netstandard

* Fix restore sources for core NuGet

* Fix NuGet packaging paths

* fix runtime package warning

* Quote restore source

* Fix runtime package paths

* Flatten runtime package library search

* Trigger build on changes

* Restore nightly workflow triggers

* FIX: dotnet restore command

Missing double quotes and https source should be first.

* TWEAK: Temporay disable workflow

We need to disable automatic trigger for the build-nuget workflow,
until we fix the issue with proper build_suffix.

* TWEAK: Extend gitignore for nuget/obj

* Fix build suffix handling

* TWEAK: Restore nuget build trigger

We have fixed the issue with build_suffixes,
so we can trigger nuget build in case something in the
nuget/cmake folder has changed.

* FIX: Build suffix issue

* FIX: NuGet restore location

* FIX: NuGet package location

There is a complex issue with picking up correct packages after build.
MSVC has a different structure than ninja and we need to merge them.
For the time being packaging is done through CI, so we can rely on this.

Additionally I have removed vanillapdf.lib, which was used to aid
semi-dynamic linking, when the package was consumed by native project.
The issues keeping up the native targeting are high and for this reason
we now rely mainly on the github clone and vcpkg.

* FIX: Build suffix naming

* FIX: Binary artifacts location

* TWEAK: Unify artifact names

* FIX: dotnet restore

Using multiple sources local and nuget.org is confusing the engine.
We can workaround this using nuget.config, where we specify the details.
Tested locally.