Fix release scripts (#56) (#57)
* FIX: GitHub release

Include code checkout, so that we can tag the current commit.
Without this there is no .git folder.

* FIX: nupkg location

In the publish nuget packages we have not updated the nupkg location.

* FIX: Cannot merge PR

PR cannot be merges into release branch if the CodeQL was not run.

Let's include this in the conditions,
that it runs on PR agains main as well as release/* branches.

* FIX: yml syntax error

* FIX: yml syntax error

* FIX: yml syntax error

* TWEAK: Codeql concurrency

When doing rapid changes on a branch the CodeQL is ran multiple times.
Let's cancel the older runs if they are still active and keep just the latest.