FIX: NuGet publishing issue
When consuming nuget package everything seems to be working.
The issue was only present when publishing the project,
where the destination of libvanillapdf.dll was in the root folder.
This cannot be picked up by vanillapdf.net, so it was not functional.
The issue comes from the migration from nuspec to csproj.

We will need to include regression tests cases,
so this does not happen in the future.
The standard build and test is not enough,
we need to include publishing as well.