FIX: Content is being included in the core package
When building the nightly package I have observed,
that all native dependencies were included in the vanillapdf core package.
The reason for this seems to be using Content within vanillapdf_net.targets.
Locally it does not happen, however I am using dotnet 9.0.301, which could be the difference.
Let's try it like this.