Ensure native runtime libraries copy on publish (#69)
* remove old ItemGroup from targets

* FIX: NuGet publishing issue

When consuming nuget package everything seems to be working.
The issue was only present when publishing the project,
where the destination of libvanillapdf.dll was in the root folder.
This cannot be picked up by vanillapdf.net, so it was not functional.
The issue comes from the migration from nuspec to csproj.

We will need to include regression tests cases,
so this does not happen in the future.
The standard build and test is not enough,
we need to include publishing as well.

* TWEAK: Drop native nuget support

Native c/c++ should be included using vcpkg.

* FIX: Content is being included in the core package

When building the nightly package I have observed,
that all native dependencies were included in the vanillapdf core package.
The reason for this seems to be using Content within vanillapdf_net.targets.
Locally it does not happen, however I am using dotnet 9.0.301, which could be the difference.
Let's try it like this.

* TWEAK: Treat nuget warnings as errors

There are sometimes warnings that nuget is not adding a file,
since the file is already there.
This is most of the time a problem even when the nuget is built.
Let's try to catch all of these as soon as possible.

(cherry picked from commit 9e1b0ba3fb83cbde8a6e9895c44c015e642935d9)